refactor(delete): split delete and deleteAll into separate files
