+++ pkg/delete/delete.go
@@ -116,62 +116,3 @@ func deleteConfig(ctx context.Context, clients client.ClientSet, t string, entri
 	}
 	return nil
 }
-
-// All collects and deletes ALL configuration objects using the provided ClientSet.
-// To delete specific configurations use Configs instead!
-//
-// Parameters:
-//   - ctx (context.Context): The context in which the function operates.
-//   - clients (ClientSet): A set of API clients used to collect and delete configurations from an environment.
-func All(ctx context.Context, clients client.ClientSet, apis api.APIs) error {
-	errCount := 0
-
-	if clients.ConfigClient == nil {
-		log.Warn("Skipped deletion of classic configurations as API client was unavailable.")
-	} else if err := classic.DeleteAll(ctx, clients.ConfigClient, apis); err != nil {
-		log.Error("Failed to delete all classic API configurations: %v", err)
-		errCount++
-	}
-
-	if clients.SettingsClient == nil {
-		log.Warn("Skipped deletion of settings configurations as API client was unavailable.")
-	} else if err := setting.DeleteAll(ctx, clients.SettingsClient); err != nil {
-		log.Error("Failed to delete all Settings 2.0 objects: %v", err)
-		errCount++
-	}
-
-	if clients.AutClient == nil {
-		log.Warn("Skipped deletion of Automation configurations as API client was unavailable.")
-	} else if err := automation.DeleteAll(ctx, clients.AutClient); err != nil {
-		log.Error("Failed to delete all Automation configurations: %v", err)
-		errCount++
-	}
-
-	if clients.BucketClient == nil {
-		log.Warn("Skipped deletion of Grail Bucket configurations as API client was unavailable.")
-	} else if err := bucket.DeleteAll(ctx, clients.BucketClient); err != nil {
-		log.Error("Failed to delete all Grail Bucket configurations: %v", err)
-		errCount++
-	}
-
-	if clients.DocumentClient == nil {
-		log.Warn("Skipped deletion of Documents configurations as appropriate client was unavailable.")
-	} else if err := document.DeleteAll(ctx, clients.DocumentClient); err != nil {
-		log.Error("Failed to delete all Document configurations: %v", err)
-		errCount++
-	}
-
-	if featureflags.Segments.Enabled() {
-		if clients.SegmentClient == nil {
-			log.Warn("Skipped deletion of %s configurations as appropriate client was unavailable.", config.SegmentID)
-		} else if err := segment.DeleteAll(ctx, clients.SegmentClient); err != nil {
-			log.Error("Failed to delete all %s configurations: %v", config.SegmentID, err)
-			errCount++
-		}
-	}
-
-	if errCount > 0 {
-		return fmt.Errorf("failed to delete all configurations for %d types", errCount)
-	}
-	return nil
-}
